+++ src/main/java/com/tertei/sorts/impl/MergeSort.java
@@ -31,14 +31,14 @@ public class MergeSort implements Sortable {
      * Rearranges the array in ascending order, using the natural order.
      * @param a the array to be sorted
      */
-    public void sort(Comparable[] a){
+    public static void sort(Comparable[] a){
         Comparable[] aux = new Comparable[a.length];
         sort(a, aux, 0, a.length - 1);
     }
 
     // recursive function
     // mergesort a[lo..hi] using auxiliary array aux[lo..hi]
-    private void sort(Comparable[] a, Comparable[] aux, int lo, int hi){
+    private static void sort(Comparable[] a, Comparable[] aux, int lo, int hi){
         if(hi <= lo) return;
         int mid = lo + (hi - lo) / 2;
         sort(a, aux, lo, mid);
@@ -48,7 +48,7 @@ public class MergeSort implements Sortable {
 
     // merge a[lo .. mid] with a[mid+1 ..hi] using aux[lo .. hi]
     // precondition: a[lo .. mid] and a[mid+1 .. hi] are sorted subarrays
-    private void merge(Comparable[] a, Comparable[] aux, int lo, int mid, int hi){
+    private static void merge(Comparable[] a, Comparable[] aux, int lo, int mid, int hi){
         // copy to auxiliary array
         for(int k = lo; k <= hi; k++){
             aux[k] = a[k];