+++ src/main/java/com/tertei/sorts/impl/QuickSort.java
@@ -25,13 +25,13 @@ public class QuickSort implements Sortable {
      * Rearranges the array in ascending order, using the natural order.
      * @param a the array to be sorted
      */
-    public void sort(Comparable[] a){
+    public static void sort(Comparable[] a){
         shuffle(a);
         sort(a, 0, a.length - 1);
     }
 
     // quicksort the subarray from a[lo] to a[hi]
-    private void sort(Comparable[] a, int lo, int hi){
+    private static void sort(Comparable[] a, int lo, int hi){
         if(hi <= lo) return;
         int j = partition(a, lo, hi);
         sort(a, lo, j-1);
@@ -40,7 +40,7 @@ public class QuickSort implements Sortable {
 
     // partition the subarray a[lo..hi] so that a[lo..j-1] <= a[j] <= a[j+1..hi]
     // and return the index j.
-    private int partition(Comparable[] a, int lo, int hi){
+    private static int partition(Comparable[] a, int lo, int hi){
         int i = lo, j = hi + 1;
         while(true){
             while(less(a[++i], a[lo]))