+++ src/main/java/com/tertei/sorts/impl/SortHelpers.java
@@ -16,18 +16,18 @@ package com.tertei.sorts.impl;
 public class SortHelpers {
 
     // is v < w ?
-    static boolean less(Comparable v, Comparable w){
+    protected static boolean less(Comparable v, Comparable w){
         return v.compareTo(w) < 0;
     }
 
     // exchange a[i] and a[j]
-    static void exch(Comparable[] a, int i, int j){
+    protected static void exch(Comparable[] a, int i, int j){
         Comparable swap = a[i];
         a[i] = a[j];
         a[j] = swap;
     }
 
-    static void shuffle(Comparable[] a){
+    protected static void shuffle(Comparable[] a){
         int N = a.length;
         for (int i = 0; i < N; i++){
             int r = (int)((i + 1.0) * Math.random());